+++ src/ila/include_all.h
@@ -17,6 +17,8 @@
 #include "ila/event/minecraft/world/actor/ArmorStandSwapItemEvent.h"
 #include "ila/event/minecraft/world/actor/DeathMessageEvent.h"
 #include "ila/event/minecraft/world/actor/DragonRespawnEvent.h"
+#include "ila/event/minecraft/world/actor/EndermanLeaveBlockEvent.h"
+#include "ila/event/minecraft/world/actor/EndermanTakeBlockEvent.h"
 #include "ila/event/minecraft/world/actor/item/ItemActorEvent.h"
 #include "ila/event/minecraft/world/actor/MobHurtEffectEvent.h"
 #include "ila/event/minecraft/world/actor/player/PlayerAteEvent.h"